+++ src/pcap_win32.cpp
@@ -1,0 +1,6 @@
+#include <pcap/pcap.h>
+#include <string>
+const char *pcap_statustostr(int) {
+	static std::string message("unknown.");
+	return message.c_str();
+}