+++ src/pktminerg.cpp
@@ -1,10 +1,15 @@
 #include <iostream>
 #include <csignal>
 #include <boost/program_options.hpp>
+#include <boost/algorithm/string.hpp>
 #include "pcaphandler.h"
 #include "socketgre.h"
+#include "socketzmq.h"
 #include "versioninfo.h"
 #include "syshelp.h"
+#ifndef WIN32
+    #include "agent_control_plane.h"
+#endif
 
 std::shared_ptr<PcapHandler> handler = nullptr;
 
@@ -18,9 +23,18 @@ int main(int argc, const char* argv[]) {
     desc.add_options()
             ("interface,i", boost::program_options::value<std::string>()->value_name("NIC"),
              "interface to capture packets")
+            ("bind_device,B", boost::program_options::value<std::string>()->value_name("BIND"),
+             "send GRE packets from this binded device.(Not available on Windows)")
+            ("pmtudisc_option,M", boost::program_options::value<std::string>()->value_name("MTU"),
+             " Select Path MTU Discovery strategy.  pmtudisc_option may be either do (prohibit fragmentation, even local one), want (do PMTU discovery, fragment locally when packet size is large), or dont (do not set DF flag)")
             ("pcapfile,f", boost::program_options::value<std::string>()->value_name("PATH"),
              "specify pcap file for offline mode, mostly for test")
-            ("remoteip,r", boost::program_options::value<std::string>()->value_name("IP"), "set gre remote ip")
+            ("remoteip,r", boost::program_options::value<std::string>()->value_name("IPs"),
+             "set gre remote IPs, seperate by ',' Example: -r 8.8.4.4,8.8.8.8")
+            ("zmq_port,z", boost::program_options::value<int>()->default_value(0)->value_name("ZMQ_PORT"),
+             "set remote zeromq server port to receive packets reliably; ZMQ_PORT default value 0 means disable.")
+            ("zmq_hwm,m", boost::program_options::value<int>()->default_value(100)->value_name("ZMQ_HWM"),
+             "set zeromq queue high watermark; ZMQ_HWM default value 100.")
             ("keybit,k", boost::program_options::value<int>()->default_value(1)->value_name("BIT"),
              "set gre key bit; BIT defaults 1")
             ("snaplen,s", boost::program_options::value<int>()->default_value(2048)->value_name("LENGTH"),
@@ -36,8 +50,12 @@ int main(int argc, const char* argv[]) {
             ("expression", boost::program_options::value<std::vector<std::string>>()->value_name("FILTER"),
              R"(filter packets with FILTER; FILTER as same as tcpdump BPF expression syntax)")
             ("dump", "specify dump file, mostly for integrated test")
+            ("control", boost::program_options::value<int>()->value_name("CONTROL_PORT"),
+             "set zmq listen port for agent daemon control. Control server won't be up if this option is not set")
             ("nofilter",
-             "force no filter; only use when you confirm that the snoop interface is different from the gre interface");
+             "force no filter; In online mode, only use when GRE interface "
+                 "is set via CLI, AND you confirm that the snoop interface is "
+                 "different from the gre interface.");
 
     boost::program_options::positional_options_description position;
     position.add("expression", -1);
@@ -76,6 +94,40 @@ int main(int argc, const char* argv[]) {
         return 1;
     }
 
+    std::string bind_device = "";
+    if (vm.count("bind_device")) {
+        bind_device = vm["bind_device"].as<std::string>();
+    }
+
+    int pmtudisc = -1;
+    int update_status = 0;
+#ifdef WIN32
+    //TODO: support pmtudisc_option on WIN32
+#else
+    if (vm.count("pmtudisc_option")) {
+        const auto pmtudisc_option = vm["pmtudisc_option"].as<std::string>();
+        if (pmtudisc_option == "do") {
+            pmtudisc = IP_PMTUDISC_DO;
+        } else if (pmtudisc_option == "dont") {
+            pmtudisc = IP_PMTUDISC_DONT;
+        } else if (pmtudisc_option == "want") {
+            pmtudisc = IP_PMTUDISC_WANT;
+        } else {
+            std::cerr << StatisLogContext::getTimeString()
+                      << "Wrong value for -M: do, dont, want are valid ones." << std::endl;
+            return 1;
+        }
+    }
+    std::shared_ptr<AgentControlPlane> agent_control_plane;
+    if (vm.count("control")) {
+        const auto daemon_zmq_port = vm["control"].as<int>();
+        agent_control_plane = std::make_shared<AgentControlPlane>(daemon_zmq_port);
+        agent_control_plane->init_msg_server();
+        update_status = 1;    
+    }
+#endif // WIN32
+
+
     if (!vm.count("remoteip")) {
         std::cerr << StatisLogContext::getTimeString() << "Please set gre remote ip with --remoteip or -r."
                   << std::endl;
@@ -83,6 +135,12 @@ int main(int argc, const char* argv[]) {
     }
 
     std::string remoteip = vm["remoteip"].as<std::string>();
+    std::vector<std::string> remoteips;
+    boost::algorithm::split(remoteips, remoteip, boost::algorithm::is_any_of(","));
+
+    int zmq_port = vm["zmq_port"].as<int>();
+    int zmq_hwm = vm["zmq_hwm"].as<int>();
+
     int keybit = vm["keybit"].as<int>();
 
     std::string filter = "";
@@ -96,16 +154,33 @@ int main(int argc, const char* argv[]) {
     bool nofilter = false;
     if (vm.count("nofilter")) {
         nofilter = true;
+        if (vm.count("interface")) {
+            if (bind_device == "") {
+                std::cerr << StatisLogContext::getTimeString() << "Can't enable --nofilter option "
+                << "because GRE bind devices(-B) is not set, GRE packet might be sent via packet captured interface(-i)"
+                << std::endl;
+                return 1;
+            } else if (bind_device == vm["interface"].as<std::string>()) {
+                std::cerr << StatisLogContext::getTimeString() << "Can't enable --nofilter option "
+                << "because packet captured interface(-i) is equal to GRE bind devices(-B)"
+                << std::endl;
+                return 1;
+            } else {
+                // valid
+            }
+        }
     }
 
-    if (!nofilter) {
-        if (filter.length() > 0) {
-            filter = filter + "and not host " + remoteip;
-        } else {
-            filter = "not host " + remoteip;
-        }
-    } else {
+    if (nofilter) {
         filter = "";
+    } else {
+        for (size_t i = 0; i < remoteips.size(); ++i) {
+            if (filter.length() > 0) {
+                filter = filter + " and not host " + remoteips[i];
+            } else {
+                filter = "not host " + remoteips[i];
+            }
+        }
     }
 
     // dump option
@@ -120,6 +195,7 @@ int main(int argc, const char* argv[]) {
     param.snaplen = vm["snaplen"].as<int>();
     param.promisc = 0;
     param.timeout = vm["timeout"].as<int>() * 1000;
+    param.need_update_status = update_status;
     int nCount = vm["count"].as<int>();
     if (nCount < 0) {
         nCount = 0;
@@ -178,10 +254,27 @@ int main(int argc, const char* argv[]) {
         }
     });
 
-    // export gre
-    std::shared_ptr<PcapExportBase> greExport = std::make_shared<PcapExportGre>(remoteip, keybit);
-    greExport->initExport();
-    handler->addExport(greExport);
+    std::shared_ptr<PcapExportBase> exportPtr = nullptr;
+    if (zmq_port != 0) {
+        exportPtr = std::make_shared<PcapExportZMQ>(remoteips, zmq_port, zmq_hwm, keybit, bind_device, param.buffer_size);
+        int err = exportPtr->initExport();
+        if (err != 0) {
+            std::cerr << StatisLogContext::getTimeString()
+                      << "zmqExport initExport failed." << std::endl;
+            return err;
+        }
+    } else {
+        // export gre
+        exportPtr = std::make_shared<PcapExportGre>(remoteips, keybit, bind_device, pmtudisc);
+        int err = exportPtr->initExport();
+        if (err != 0) {
+            std::cerr << StatisLogContext::getTimeString()
+                      << "greExport initExport failed." << std::endl;
+            return err;
+        }
+    }
+    handler->addExport(exportPtr);
+
 
     // begin pcap snoop
 
@@ -190,6 +283,6 @@ int main(int argc, const char* argv[]) {
     std::cout << StatisLogContext::getTimeString() << "End pcap snoop." << std::endl;
 
     // end
-    greExport->closeExport();
+    exportPtr->closeExport();
     return 0;
 }