+++ src/syshelp.h
@@ -5,6 +5,6 @@
 int set_high_setpriority();
 
 // bind cpu to given core, return 0 if success
-int set_cpu_affinity(int cpu);
+int set_cpu_affinity(int);
 
 #endif // SRC_SYSHELP_H_